Dropbox 4

+++ app/Http/Controllers/FilesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\File;
 //use DB; Enkel bij gebruik SQL queries ipv Eloquent (Zie index()->DB)
 
@@ -16,25 +17,6 @@ class FilesController extends Controller
     public function __construct()
     {
         $this->middleware('auth', ['except' => ['index', 'show']]);
-    }
-    
-    
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        // $files = File::all();
-        // $files = File::orderBy('filename', 'desc')->get(); 
-        // $files = File::orderBy('filename', 'desc')->take(1)->get();  // take(limit)
-        // $files = File::where('filename', 'Tweede')->get(); 
-        // $files = DB::select('SELECT * from files where filename = "Eerste"'); // SQL query (use DB) 
-        // $files = File::orderBy('filename', 'desc')->simplePaginate(6); // Pagina's met enkel vorige/volgende knoppen
-           $files = File::orderBy('created_at', 'desc')->paginate(5); // Pagina's met nummers
-
-        return view('files.index')->with('files', $files);
     }
 
     /**
@@ -57,13 +39,38 @@ class FilesController extends Controller
     {
         $this->validate($request, [
             'filename' => 'required',
-//            'file' => 'required'
+            'bestand' => 'required|max:1999' // Want apache default upload size = 2MB
         ]);
+        
+        // Handle file upload
+        
+        if($request->hasFile('bestand')){
+            
+            // Get filename with the extension
+            $filenameWithExt = $request->file('bestand')->getClientOriginalName();
+            
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            
+            // Get just extension
+            $extension = $request->file('bestand')->getClientOriginalExtension();
+            
+            // Filename to store
+            $fileNameToStore = $filename . '_' . time() . '.' .$extension;
+            
+            // Get file size
+            $filesize = $request->file('bestand')->getClientSize();
+            
+            // Upload file
+            $path = $request->file('bestand')->storeAs('public/bestanden', $fileNameToStore);
+        }
        
         // Create file
         $file = new File;
         $file->filename = $request->input('filename');
         $file->user_id = auth()->user()->id;
+        $file->size = $filesize;
+        $file->bestand = $fileNameToStore;
         $file->save();
         
         return redirect('/dashboard')->with('success', 'Bestand toegevoegd');
@@ -91,6 +98,16 @@ class FilesController extends Controller
     public function edit($id)
     {
         $file =  File::find($id);
+        
+        // Check for correct user
+        if(auth()->user()->id !== $file->user_id){
+            return redirect('/dashboard')->with('error', 'Niet toegelaten');
+        }
+//         if ($request->hasFile('bestand')) {
+//            Storage::delete('public/bestanden/' . $file->bestand);
+//            $file->bestand = $fileNameToStore;
+//        }﻿
+        
         return view('files.edit')->with('file', $file);
 
     }
@@ -105,13 +122,42 @@ class FilesController extends Controller
     public function update(Request $request, $id)
     {
         $this->validate($request, [
-            'filename' => 'required',
-//          'file' => 'required'
+            'filename' => 'required'
         ]);
+        
+        
+            // Handle file upload       
+            if($request->hasFile('bestand')){
+                
+                // Get filename with the extension
+                $filenameWithExt = $request->file('bestand')->getClientOriginalName();
+
+                // Get just filename
+                $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+                // Get just extension
+                $extension = $request->file('bestand')->getClientOriginalExtension();
+
+                // Filename to store
+                $fileNameToStore = $filename . '_' . time() . '.' .$extension;
+
+                // Upload file
+                $path = $request->file('bestand')->storeAs('public/bestanden', $fileNameToStore);
+        }
        
         // Create file
         $file = File::find($id);
+        
+         // Check for correct user
+        if(auth()->user()->id !== $file->user_id){
+            return redirect('/dashboard')->with('error', 'Niet toegelaten');
+        }
+        
         $file->filename = $request->input('filename');
+        if($request->hasFile('bestand')){
+                $file->bestand = $fileNameToStore;
+        }
+
         $file->save();
         
         return redirect('/dashboard')->with('success', 'Bestand aangepast');
@@ -126,8 +172,15 @@ class FilesController extends Controller
     public function destroy($id)
     {
         $file = File::find($id);
-        $file->delete();
         
+        // Check for correct user
+        if(auth()->user()->id !== $file->user_id){
+            return redirect('/dashboard')->with('error', 'Niet toegelaten');
+        }
+        
+        Storage::delete('public/bestanden/' . $file->bestand);
+        
+        $file->delete();
         return redirect('/dashboard')->with('success', 'Bestand verwijderd');
     }
 }
